refactor: organize filtering methods and display better
